Evan's big fix
This is a huge commit. It includes creating a respawn controller, fixing the pause controller, modifying the player controller to work with the camera system, respawn system, and pause system, and finally incorporating all of that into a scene template for people to make levels off of. Files were also slightly reorganized to include folders for various supporting assets, putting the prefabs front and center.